Renaming files
- all moved into exodusIIcpp directory
- using snake_case for file names

Closes #60
