--- include/exodusIIcpp.h
@@ -1,13 +1,0 @@
-// SPDX-FileCopyrightText: 2022 David Andrs <andrsd@gmail.com>
-// SPDX-License-Identifier: MIT
-
-#pragma once
-
-#include "exodusIIcppElementBlock.h"
-#include "exodusIIcppEnums.h"
-#include "exodusIIcppError.h"
-#include "exodusIIcppException.h"
-#include "exodusIIcppFile.h"
-#include "exodusIIcppFilesystem.h"
-#include "exodusIIcppNodeSet.h"
-#include "exodusIIcppSideSet.h"
