+++ src/exception.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 David Andrs <andrsd@gmail.com>
 // SPDX-License-Identifier: MIT
 
-#include "exodusIIcppException.h"
+#include "exodusIIcpp/exception.h"
 
 namespace exodusIIcpp {
