+++ src/node_set.cpp
@@ -1,8 +1,8 @@
 // SPDX-FileCopyrightText: 2022 David Andrs <andrsd@gmail.com>
 // SPDX-License-Identifier: MIT
 
-#include "exodusIIcppNodeSet.h"
-#include "exodusIIcppException.h"
+#include "exodusIIcpp/node_set.h"
+#include "exodusIIcpp/exception.h"
 
 namespace exodusIIcpp {
