+++ test/Error_test.cpp
@@ -1,5 +1,5 @@
 #include "gmock/gmock.h"
-#include "exodusIIcpp.h"
+#include "exodusIIcpp/exodusIIcpp.h"
 
 using namespace exodusIIcpp;
